Fixed issue with fetching payment_method_id
This commit fixes a potential bug in fetching payment_method_id for payment params for an order.

Earlier we were using a ActiveRecord query to fetch the payment_method_id but this could have been a potential bug so this has now been changed. Now, payment_method_id is fetched from the options selected by the user from the payment form.